pyproject.toml
[project]
name = "course-materials-rag-system"
version = "0.1.0"
description = "A Retrieval-Augmented Generation system for answering questions about course materials."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb==1.0.15",
    "zai-sdk>=0.2.2",
    "sentence-transformers==5.0.0",
    "fastapi==0.116.1",
    "uvicorn==0.35.0",
    "python-multipart==0.0.20",
    "python-dotenv==1.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[tool.pytest.ini_options]
pythonpath = [
    ".",
]
